CHANGELOG: Dividing two ints should return a float64, not a float32 (#10553)
* CHANGELOG: Dividing two ints should return a float64, not a float32

* Interpreter also needed to handle division correctly

* Fixed definition of BinaryOp.emit to correctly emit F64 when doing integer division

* Fixed up many cases in the division test

* Fixed expression type calculation for dividing int types

* Fixed the type in roc_curve

* Fixed a test that still expected floats

+++ hail/src/main/scala/is/hail/expr/ir/Interpret.scala
@@ -133,7 +133,7 @@ object Interpret {
                 case Add() => ll + rr
                 case Subtract() => ll - rr
                 case Multiply() => ll * rr
-                case FloatingPointDivide() => ll.toFloat / rr.toFloat
+                case FloatingPointDivide() => ll.toDouble / rr.toDouble
                 case RoundToNegInfDivide() => java.lang.Math.floorDiv(ll, rr)
                 case BitAnd() => ll & rr
                 case BitOr() => ll | rr
@@ -157,7 +157,7 @@ object Interpret {
                 case Add() => ll + rr
                 case Subtract() => ll - rr
                 case Multiply() => ll * rr
-                case FloatingPointDivide() => ll.toFloat / rr.toFloat
+                case FloatingPointDivide() => ll.toDouble / rr.toDouble
                 case RoundToNegInfDivide() => java.lang.Math.floorDiv(ll, rr)
                 case BitAnd() => ll & rr
                 case BitOr() => ll | rr
